Use consistent graph size for cubism's sake
Change the step period to always be 600, which is only a small adjustment. Set
cubism to assume this number for both the total period, and width of the graph,
so it's happy. Rickshaw doesn't mind either way, because it can be explicitly
told the time, as opposed to cubism needing the size and step and the width of
the div to match