Prevent color race by pre-creating color palette array
Normally the order of palette color allocation would depend on which json call
returned first. However, this means that the color sequence down the page is
not always the same.

Instead: allocate a color array and use the color[index] for the graph of the
same index. Ensures the color order is consistent with the palette.