Try to get metrics before searching
Attempt to get the metrics list in order to force an invokation of any backend
errors that will be captured. Return is an array of all metrics in cases of
redis caching, but it will at least attempt to fire any network issues before
attempting to search for previously cached source lists that won't return any
data anyway if there are network issues.