Set all bases to 1000 - bug introducing - pending logic fix
We can't always guarantee that a unit will be able to be shown in orders of
magnitude in 1024 bases. If we have time, or MB or generic units (LOC?) we
can't be sure we're showing the right data units.

SO

For now, we'll throw everything to base 1000.

This is what horizon was showing anyway, so there was a disconnect between
graph types.