Correct issues with nose dive data, NaN data on stacked graphs
For a metric with σ = 0, it would appear as having no y-axis. This is corrected.
For a metric with σ = 0 and x̄ = 0, it would appear as having no value nor axis.
This is corrected.

Logic regarding how to handle a lack of data in descartes dataset padding is
corrected, and handled in rickshaw: the nil value is used in the controller,
which is explicitly handled in rickshaw to cast back to a NaN