Use calculated height, as opposed to hardcoded height, for window
sb-admin uses hardcoded values, so at smaller heights, there's excess
whitespace.

Using 100vh uses the 'viewport height' as the base, which dynamically changes
with rescaling. Removing the 52px ensures that the header height is removed
from the calculation.

Result: any resizing has the no scrollbar until the height is sufficently short
to not be able to show all elements on the page.