Zoom to Selected functionality
For graphs with range sliders, when the window.location.hash is changed, create
a link to set the bounds of the range slider to the start and stop of the url.
This will allow the selected area to be re-queried, usually at a higher
resolution.

Also added check to ensure time period less than 600 is not inadvetently
selected.