Add a Backend Library Description option for Settings
Useful to be able to define something as being related to something as
opposed to it's library name (e.g. sieste <-> vaultaire)

Defaults to library name, but allows for library-based overloading

Doesn't require working settings, just a good title to call the right
backend description method. If it can't be found, it defaults to the
value given for "name".

This should be the smallest profile plumbing to get "Vaultaire"
displaying in place of Sieste on the settings modal