Turn select2 in-input into table
instead of making li pills, add them to a table as divs. keep the classes the same because select2 creates events based on these. move the select dropdown to a col6 and the selected to another col6. force modal size.

very VERY much a quick friday arvo hack, work in progress