Retroactively remove metric url content from buttons
The button's urls are created at page load, which include references to
metric count = n (n -1 in cases of sidebar remove buttons). When n
reduces to n-1, the number of metrics in all these buttons also needs to
change. So, take all urls on the page containing the metric name and
remove it from their href query string.