update deployment process
settings modal takes the current git version, OR the .gvb_version file.
BUT if neither exist, it doesn't show anything.

The deployment process creates the .gvb_version file as a cat of `git
version-bump show` and commits it to the temporary branch (which is only
used by giddyup, and never goes to github)

This should match with the commentary from
mpalmer/git-version-bump:issue#2