init commit for BOM code
takes a folder full of files that are a JSON blob of the Bureau of Meterology Observations (found on pages like http://www.bom.gov.au/products/IDN60901/IDN60901.94759.shtml, bottom of the page, 'JavaScript Object Notation format (JSON) in row-major order'.

IDN is the store, BOM is the source.