Incorporate clizia's changes for accepting datafeeds
Clizia does not do the logic for feeds, it now only accepts an array of datapoints, or a feed. Machiavelli now has to handle the feed creation.

The update command in Rickshaw graphs takes an object that includes the one or more metric objects under the metric key. The update loops for standard and stacked graphs have been altered accordingly.